refactor(office-preview): move previews under main features
