+++ src/main/features/officePreview/excel/chart/__tests__/excelChartRenderer.test.ts
