+++ src/main/features/officePreview/excel/chart/excelChartRenderer.ts
