+++ src/main/features/officePreview/excel/internal/excelPreviewUtils.ts
